Organize cpp_audio directory
