+++ src/cpp_audio/core/VarUtils.h
