+++ src/cpp_audio/synths/NoiseFlanger.cpp
@@ -1,6 +1,6 @@
 #include "NoiseFlanger.h"
 #include "AudioUtils.h"
-#include "../VarUtils.h"
+#include "../core/VarUtils.h"
 #include <juce_dsp/juce_dsp.h>
 #include <vector>
 #include <algorithm>
