+++ src/cpp_audio/synths/SynthFunctions.h
