+++ src/cpp_audio/ui/StepListPanel.h
@@ -3,7 +3,7 @@
 #include <juce_gui_basics/juce_gui_basics.h>
 #include <vector>
 #include "StepConfigPanel.h"
-#include "../Track.h"
+#include "../core/Track.h"
 
 class StepListPanel : public juce::Component,
                       private juce::ListBoxModel,
