+++ src/cpp_audio/ui/SubliminalDialog.cpp
@@ -1,7 +1,7 @@
 // SubliminalDialog C++ implementation.
 // Translated from src/audio/ui/subliminal_dialog.py and adapted for JUCE.
 
-#include "../cpp_audio/Track.h"
+#include "../cpp_audio/core/Track.h"
 #include "SubliminalDialog.h"
 #include <juce_gui_basics/juce_gui_basics.h>
 #include <cmath>
